cafViewer: Support high dpi for overlay image
Set the size of m_overlayPaintingQImage to the widget size times the
pixel ratio and set the pixel ratio to the same as the widget.

QImage operates in actual pixels and not device independent pixels.
Without multiplying with the pixel ratio the overlay image was only
covering the lower left quarter of the widget when using a high dpi
display with pixel ratio of 2.

Setting the pixel ratio for the QImage ensures that text is scaled
correctly using device independent pixel/font size.

This is only relevant if Qt::AA_EnableHighDpiScaling is set. If not the
pixel ratio will always return 1 and work exactly as before.